Cart added

+++ MusicalShop.Services/ProductService.cs
@@ -6,6 +6,7 @@
     using MusicalShop.Data.Models;
     using MusicalShop.Services.Mapping;
     using MusicalShop.Services.Models;
+    using System;
     using System.Linq;
     using System.Threading.Tasks;
     public class ProductService : IProductService
@@ -19,7 +20,17 @@
 
         public async Task<bool> CreateProductAsync(ProductServiceModel model)
         {
+            var productTypeFromDb = await this.context.ProductTypes
+                .SingleOrDefaultAsync(productType => productType.Name == model.ProductType.Name);
+
+            if (productTypeFromDb == null)
+            {
+                throw new ArgumentNullException(nameof(productTypeFromDb));
+            }
+
             var product = Mapper.Map<Product>(model);
+            product.Id = Guid.NewGuid().ToString();
+            product.ProductType = productTypeFromDb;
 
             await context.Products.AddAsync(product);
             var result = await context.SaveChangesAsync();
@@ -40,6 +51,11 @@
         {
             var product = await context.Products.SingleOrDefaultAsync(x => x.Id == id);
 
+            if (product == null)
+            {
+                throw new ArgumentNullException(nameof(product));
+            }
+
             context.Products.Remove(product);
             var result = await context.SaveChangesAsync();
             return result > 0;
@@ -47,18 +63,28 @@
 
         public async Task<bool> EditProductAsync(string id, ProductServiceModel model)
         {
-            var productypeFromDb = await context.ProductTypes.FirstOrDefaultAsync(x => x.Name == model.ProductType.Name);
+            var producTypeFromDb = await context.ProductTypes.FirstOrDefaultAsync(x => x.Name == model.ProductType.Name);
 
+            if (producTypeFromDb == null)
+            {
+                throw new ArgumentNullException(nameof(producTypeFromDb));
+            }
 
             var productFromDb = await context.Products.SingleOrDefaultAsync(x => x.Id == id);
+
+            if (productFromDb == null)
+            {
+                throw new ArgumentNullException(nameof(productFromDb));
+            }
 
             if (model.Picture != null)
             {
                 productFromDb.Picture = model.Picture;
             }
+
             productFromDb.Name = model.Name;
             productFromDb.Price = model.Price;
-            productFromDb.ProductType = productypeFromDb;
+            productFromDb.ProductType = producTypeFromDb;
             productFromDb.Quantity = model.Quantity;
             productFromDb.ManufacturedOn = model.ManufacturedOn;
             productFromDb.Description = model.Description;
